fix up playback of records
Previously the 'tm_isdst' setting was free-flowing, which could
result in playback errors (e.g. seen recorded timestamp shifting
by an hour, but only on the first record, which would result in
"hung"-looking playback while it is waiting for an hour.

Explicit setting of this value first should make things deterministic.